requirements.txt
black~=21.6b0
mypy~=0.902
pytest~=6.2.4
python-telegram-bot~=13.6
natasha~=1.4.0
requests~=2.25.1
types-requests~=0.1.11
